requirements.txt
autogen
json
matplotlib
networkx
numpy
pandas
plotly
python-dotenv